renamed component files and code styles
