+++ aim/web/ui/src/components/ImagePropertiesPopover/ImagePropertiesPopover.d.ts
