+++ aim/web/ui/src/components/kit/Slider/Slider.scss
